Split sp-enhancements mono-module into separate files
