+++ src/sp-enhancements/_constants.ts
@@ -1,0 +1,12 @@
+/**
+ * @ignore
+ */
+export const CHARACTER_CODE_FOR_UPPERCASE_A = 'A'.charCodeAt(0);
+/**
+ * @ignore
+ */
+export const CHARACTER_CODE_FOR_LOWERCASE_A = 'a'.charCodeAt(0);
+/**
+ * @ignore
+ */
+export const CHARACTER_CODE_FOR_NUMBER_0 = '0'.charCodeAt(0);
