+++ src/sp-enhancements/getSpEntryWithId.ts
@@ -1,0 +1,13 @@
+import { ISpEnhancementEntry } from '../datamine-types';
+import { getSpEntryId } from '../sp-enhancements';
+
+/**
+ * @description Get the first SP Enhancement Entry that matches the given SP Entry ID, if it exists.
+ * @param id SP Enhancement entry ID
+ * @param entries Collection of SP Enhancement entries to search in
+ * @returns The corresponding SP Enhancement entry with the given SP ID, undefined otherwise.
+ */
+export default function getSpEntryWithId (id: string, entries: ISpEnhancementEntry[]): ISpEnhancementEntry | undefined {
+	const spId = getSpEntryId(id);
+	return (id && Array.isArray(entries) && entries.find(e => getSpEntryId(e && e.id) === spId)) || void 0;
+}
