+++ src/sp-enhancements/index.ts
@@ -1,0 +1,9 @@
+export { default as getEffectsForSpEnhancement } from './getEffectsForSpEnhancement';
+export { default as getSpCategoryName } from './getSpCategoryName';
+export { default as spIndexToCode } from './spIndexToCode';
+export { default as spCodeToIndex } from './spCodeToIndex';
+export { default as getSpEntryId } from './getSpEntryId';
+export { default as getSpEntryWithId } from'./getSpEntryWithId';
+export { default as getAllDependenciesForSpEntry } from './getAllDependenciesForSpEntry';
+export { default as getAllEntriesThatDependOnSpEntry } from './getAllEntriesThatDependOnSpEntry';
+
